Refactor imports and remove deprecated modules
Updated all import paths to use new subpackage structure, replacing legacy flat imports. Removed deprecated files including examples/run_gimbal_demo.py and gimbal_pymc/identifiability.py. Updated documentation and test files to reflect new organization, ensuring all modules use explicit subpackage imports for improved maintainability.
