+++ debug/triangulation/analyze_skeleton_3d.py
@@ -2,7 +2,7 @@
 
 import numpy as np
 from gimbal_pymc import DEMO_V0_1_SKELETON, SyntheticDataConfig, generate_demo_sequence
-from gimbal_pymc.camera_utils import camera_center_from_proj
+from gimbal_pymc.cameras.projection import camera_center_from_proj
 
 config = SyntheticDataConfig(T=100, C=3, S=3, kappa=5.0, random_seed=42)
 data = generate_demo_sequence(DEMO_V0_1_SKELETON, config)
