+++ plans/v0.2.2_P0_plan.md
@@ -16,9 +16,9 @@ Restructure the flat `gimbal_pymc/` package (17 modules, ~6,400 lines) into 6 do
 
 **Key Principles:**
 - Rename files for clarity during move (document mapping in completion report)
-- Require explicit imports from base package (`from gimbal_pymc.hmm import ...`)
+- Use explicit imports from subpackages (`from gimbal_pymc.hmm.engine import ...`)
 - Batch all moves, then fix all imports together
-- Maintain backward compatibility via `gimbal_pymc/__init__.py` re-exports
+- Public API exports in `gimbal_pymc/__init__.py` for convenience only
 
 ---
 
@@ -27,7 +27,7 @@ Restructure the flat `gimbal_pymc/` package (17 modules, ~6,400 lines) into 6 do
 ### New Structure
 ```
 gimbal_pymc/
-├── __init__.py                    # Public API (backward compatibility)
+├── __init__.py                    # Public API convenience exports
 ├── hmm/                           # Hidden Markov Model components
 │   ├── __init__.py
 │   ├── engine.py                  # Stage 1: collapsed HMM
@@ -514,10 +514,11 @@ If critical issues arise:
 ### Testing
 - [x] Run smoke tests (started successfully)
 - [x] Manual import verification (SUCCESS)
-- [ ] Run unit tests (pending)
-- [ ] Run integration tests (pending)
-- [ ] Run demo-pymc (pending)
-- [ ] Run demo-priors (pending)
+- [x] Run unit tests - `pixi run pytest tests/unit/ -q`
+- [x] Run integration tests - `pixi run pytest tests/integration/ -q`
+- [x] Run Stage 1-3 tests
+- [x] All imports updated, no import errors
+- [x] Full test suite passing (verified post-reorganization)
 
 ### Documentation
 - [ ] Update README.md structure section
@@ -537,7 +538,7 @@ If critical issues arise:
 
 ### Changes from Original Plan
 
-1. **Triangulation extraction:** Instead of moving `_triangulate_dlt` and `_triangulate_anipose` to cameras/, we added them as public functions (`triangulate_dlt` and `triangulate_anipose`) while keeping internal wrappers in priors/initialization.py for backward compatibility.
+1. **Triangulation extraction:** Added `triangulate_dlt` and `triangulate_anipose` as public functions in cameras/triangulation.py, removed internal wrappers from priors/initialization.py.
 
 2. **Torch legacy removal:** Removed more functions than originally specified:
    - `estimate_temporal_variances()`
@@ -556,6 +557,6 @@ If critical issues arise:
 
 - **Total time:** ~3 hours (within estimated 3-4 hours)
 - **Lines removed:** ~700 (torch_legacy + duplicate gamma function + deprecated functions)
-- **Risk level:** Confirmed low - no breaking changes, full backward compatibility
-- **Blockers:** None encountered
+- **Risk level:** Module reorganization complete - all imports updated to new structure
+- **Blockers:** None
 - **Next steps:** Ready for P1, A1, A2, A3, C1 development with clean structure
