+++ tests/diagnostics/v0_2_1_divergence/test_group_3_sampling_baseline_minimal.py
@@ -98,7 +98,7 @@ def build_model(data):
 
     # Initialize using library estimator
     print("  Initializing from observations (DLT triangulation)...")
-    from gimbal_pymc.fit_params import initialize_from_observations_dlt
+    from gimbal_pymc.priors.initialization import initialize_from_observations_dlt
 
     init_result = initialize_from_observations_dlt(
         y_observed=data["y_2d"],
