+++ tests/diagnostics/v0_2_1_divergence/test_group_7_likelihood_only_freeze_latents.py
@@ -94,7 +94,7 @@ def build_model_frozen(data):
     C, T, K, _ = data["y_2d"].shape
 
     print("  Initializing from observations (DLT triangulation)...")
-    from gimbal_pymc.fit_params import initialize_from_observations_dlt
+    from gimbal_pymc.priors.initialization import initialize_from_observations_dlt
 
     init_result = initialize_from_observations_dlt(
         y_observed=data["y_2d"],
