+++ tests/diagnostics/v0_2_1_divergence/test_group_8c_temporal_T500.py
@@ -63,7 +63,7 @@ def load_dataset(dataset_name, max_T):
 def build_model(data):
     """Build full model with mixture and HMM (matches Stage H configuration)."""
     print("  Initializing from observations (DLT triangulation)...")
-    from gimbal_pymc.fit_params import initialize_from_observations_dlt
+    from gimbal_pymc.priors.initialization import initialize_from_observations_dlt
     
     init_result = initialize_from_observations_dlt(
         y_observed=data["y_2d"],
