+++ tests/pipeline/stage_g_model_build.py
@@ -19,7 +19,8 @@ import pymc as pm
 
 # Add repo root to path
 
-import gimbal_pymc as gp 
+import gimbal_pymc as gp
+
 
 def load_all_preprocessing_outputs(dataset_dir: Path, fits_dir: Path) -> Dict[str, Any]:
     """Load outputs from all preprocessing stages."""
@@ -95,7 +96,7 @@ def run_stage_g(dataset_dir: Path, fits_dir: Path, output_dir: Path) -> Dict[str
 
     # Initialize from observations using library estimator
     print("\n[2/4] Computing initialization from observations...")
-    from gimbal_pymc.fit_params import initialize_from_observations_dlt
+    from gimbal_pymc.priors.initialization import initialize_from_observations_dlt
 
     # Use library estimator (DLT triangulation + robust estimation)
     # This ensures parents array is used consistently
