+++ tests/smoke/test_full_pipeline_smoke.py
@@ -23,7 +23,7 @@ from gimbal_pymc import (
     DEMO_V0_1_SKELETON,
     SyntheticDataConfig,
 )
-from gimbal_pymc.fit_params import initialize_from_observations_dlt
+from gimbal_pymc.priors.initialization import initialize_from_observations_dlt
 
 
 def test_v0_1_pipeline_builds():
